+++ eslint.config.js
@@ -1,0 +1,8 @@
+const toreda = require('@toreda/eslint-config');
+
+module.exports = [
+	{
+		ignores: ['dist/**', 'coverage/**', 'docs/**', 'node_modules/**']
+	},
+	...toreda
+];