+++ gulpfile.ts
@@ -1,4 +1,28 @@
-//const eslint = require('gulp-eslint');
+/**
+ *	MIT License
+ *
+ *	Copyright (c) 2019 - 2022 Toreda, Inc.
+ *
+ *	Permission is hereby granted, free of charge, to any person obtaining a copy
+ *	of this software and associated documentation files (the "Software"), to deal
+ *	in the Software without restriction, including without limitation the rights
+ *	to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
+ *	copies of the Software, and to permit persons to whom the Software is
+ *	furnished to do so, subject to the following conditions:
+
+ * 	The above copyright notice and this permission notice shall be included in all
+ * 	copies or substantial portions of the Software.
+ *
+ * 	THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+ *	IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+ *	FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+ * 	AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+ *	LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+ *	OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
+ * 	SOFTWARE.
+ *
+ */
+
 import gulp, {dest, series, src} from 'gulp';
 
 import {ESLint} from 'eslint';
@@ -10,7 +34,6 @@ const eslint = new ESLint({
 });
 
 const srcPatterns = ['src/**.ts', 'src/**/*.ts'];
-
 const tsc = ts.createProject('tsconfig.json');
 
 async function linter() {
@@ -24,15 +47,24 @@ async function linter() {
 function createDist() {
 	// Hack to create folder structures without actually reading files.
 	// Nested folders need to be created in their nested order.
-	return gulp.src('*.*', {read: false}).pipe(gulp.dest('./dist'));
+	// NOTE: In other projects this task is handled by `@toreda/build-tools`, however
+	// we cannot use it here because this package is a dependency of build-tools.
+	return src('*.*', {read: false}).pipe(gulp.dest('./dist'));
 }
 
-function cleanDist() {
-	return del(`dist/**`, {force: true});
+async function cleanDist() {
+	// NOTE: In other projects this task is handled by `@toreda/build-tools`, however
+	// we cannot use it here because this package is a dependency of build-tools.
+	return del(`dist/**`, {
+		force: true,
+		dryRun: false
+	});
 }
 
 function buildSrc() {
 	// Build typescript sources and output them in './dist'.
+	// NOTE: In other projects this task is handled by `@toreda/build-tools`, however
+	// we cannot use it here because this package is a dependency of build-tools.
 	return src(srcPatterns).pipe(tsc()).pipe(dest('dist'));
 }